refactor(desktop): D38 P2 split App.tsx into session, settings, and workspace hooks
Extract useAppSession (session/run/agent/workflow/footer wiring),
useAppSettings (theme/models/permission preferences), and
useAppWorkspace (workspace/repository/file-preview wiring).
App.tsx remains the startup/orchestration shell with MainLayout JSX.
